+++ rust/experimental/lang/manifest.toml
@@ -4,8 +4,3 @@ scope = "global"
 activations = ["autorun"]
 exec_path="./bin/xi-lang"
 commands = []
-
-[[languages]]
-name = "Rust"
-extensions = ["rs"]
-scope = "source.rust"
